pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "istampit-cli"
version = "1.0.3"
description = "iStampit CLI – Blockchain timestamp verification with OpenTimestamps"
authors = [ { name = "SinAI Inc.", email = "pypi@eihdah.com" } ]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "opentimestamps-client>=0.7.0",
]

[project.scripts]
istampit = "istampit_cli.__main__:main"

[project.urls]
Homepage = "https://github.com/SinAi-Inc/istampit-cli"
Documentation = "https://github.com/SinAi-Inc/istampit-cli#readme"
Source = "https://github.com/SinAi-Inc/istampit-cli"
Tracker = "https://github.com/SinAi-Inc/istampit-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
